Separate module for album
